+++ BetterCrafting/Patches/CraftingPage_Patches.cs
@@ -1,9 +1,17 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Emit;
 
 using HarmonyLib;
+
+using Netcode;
+
 using StardewModdingAPI;
+
+using StardewValley;
 using StardewValley.Menus;
+using StardewValley.Network;
 
 namespace Leclair.Stardew.BetterCrafting.Patches;